Correctly match exceptions received within two output callbacks
The output filter can be called right in the middle of the output of an
exception. This isn't new, but it didn't seem to have been common until
more recent versions of emacs.

Restricting the search to the last output position doesn't allow the
regex to match correctly, and exceptions get silently ignored.

When the filter is run, expand the lookback to include the necessary
lines required for a previous partial output to match.